+++ packages/pluggableWidgets/datagrid-web/src/components/WidgetFooter.tsx
@@ -3,14 +3,22 @@ import { observer } from "mobx-react-lite";
 import { ReactElement } from "react";
 import { SelectionCounter } from "../features/selection-counter/SelectionCounter";
 import { useSelectionCounterViewModel } from "../features/selection-counter/injection-hooks";
-import { useDatagridConfig, usePaginationService, useTexts } from "../model/hooks/injection-hooks";
+import {
+    useCustomPagination,
+    useDatagridConfig,
+    usePaginationConfig,
+    usePaginationVM,
+    useTexts
+} from "../model/hooks/injection-hooks";
 import { Pagination } from "./Pagination";
 
 export const WidgetFooter = observer(function WidgetFooter(): ReactElement {
     const config = useDatagridConfig();
-    const paging = usePaginationService();
+    const pgConfig = usePaginationConfig();
+    const paging = usePaginationVM();
     const { loadMoreButtonCaption } = useTexts();
     const selectionCounterVM = useSelectionCounterViewModel();
+    const customPagination = useCustomPagination();
 
     return (
         <div className="widget-datagrid-footer table-footer">
@@ -35,6 +43,7 @@ export const WidgetFooter = observer(function WidgetFooter(): ReactElement {
                     <If condition={config.pagingPosition !== "top"}>
                         <Pagination />
                     </If>
+                    <If condition={pgConfig.customPaginationEnabled}>{customPagination.get()}</If>
                 </div>
             </div>
         </div>