+++ packages/pluggableWidgets/datagrid-web/src/model/configs/Datagrid.config.ts
@@ -29,6 +29,8 @@ export interface DatagridConfig {
     columnsResizable: boolean;
     columnsSortable: boolean;
     isInteractive: boolean;
+    requestTotalCount: boolean;
+    constPageSize: number;
 }
 
 export function datagridConfig(props: DatagridContainerProps): DatagridConfig {
@@ -58,7 +60,9 @@ export function datagridConfig(props: DatagridContainerProps): DatagridConfig {
         columnsFilterable: props.columnsFilterable,
         columnsResizable: props.columnsResizable,
         columnsSortable: props.columnsSortable,
-        isInteractive: isInteractive(props)
+        isInteractive: isInteractive(props),
+        requestTotalCount: requestTotalCount(props),
+        constPageSize: props.pageSize
     };
 
     return Object.freeze(config);
@@ -94,4 +98,8 @@ function selectionType(props: DatagridContainerProps): SelectionType {
 
 function selectionMethod(props: DatagridContainerProps): SelectionMethod {
     return props.itemSelection ? props.itemSelectionMethod : "none";
+}
+
+function requestTotalCount(props: DatagridContainerProps): boolean {
+    return props.pagination === "buttons" || props.showNumberOfRows;
 }