+++ packages/shared/widget-plugin-grid/src/main.ts
@@ -8,6 +8,8 @@ export { type SelectionHelperService } from "./interfaces/SelectionHelperService
 export type { TaskProgressService } from "./interfaces/TaskProgressService";
 export { createFocusController } from "./keyboard-navigation/createFocusController";
 export { layoutAtom } from "./keyboard-navigation/layout.model";
+export { PageSizeStore } from "./pagination/PageSize.store";
+export * from "./pagination/pagination.model";
 export { SelectAllService } from "./select-all/SelectAll.service";
 export { SelectionCounterViewModel } from "./selection-counter/SelectionCounter.viewModel";
 export * from "./selection/context";