+++ packages/pluggableWidgets/datagrid-web/CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - We added a new property for export to excel. The new property allows to set the cell export type and also the format for type number and date.
 
+- We have introduced the "Page" and "Page Size" attributes to provide complete control over DataGrid pagination.
+
 ## [3.7.0] - 2025-11-11
 
 ### Added